# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atlas Backend API
  version: 0.1.0
  description: >
    OpenAPI documentation for the Atlas backend API. This spec covers the
    auth, users, voice, drafts, analytics, alerts, images, research, and
    trending routes exposed by the Express service.
servers:
  - url: /
    description: Current server
...
security:
  - bearerAuth: []
paths:
...
  /api/analytics/summary:
    get:
      tags:
        - Analytics
      summary: Get a 30-day analytics summary
      responses:
        '200':
          description: Summary metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/analytics/learning-log:
    post:
      tags:
        - Analytics
      summary: Create a learning log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLearningLogRequest'
      responses:
        '200':
          description: Entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningLogEnvelope'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Analytics
      summary: List recent learning log entries
      responses:
        '200':
          description: Learning log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningLogListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/analytics/engagement:
    get:
      tags:
        - Analytics
      summary: List recent engagement events
      responses:
        '200':
          description: Engagement events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsEventsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/analytics/engagement-daily:
    get:
      tags:
        - Analytics
      summary: Compare predicted versus actual engagement by day
      responses:
        '200':
          description: Daily engagement points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementDailyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/analytics/activity-daily:
    get:
      tags:
        - Analytics
      summary: Get a 30-day activity sparkline
      responses:
        '200':
          description: Daily activity points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDailyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/analytics/team-engagement-daily:
    get:
      tags:
        - Analytics
      summary: Get daily team engagement metrics
      responses:
        '200':
          description: Team engagement points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamEngagementDailyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/analytics/days-to-peak:
    get:
      tags:
        - Analytics
      summary: Measure days-to-peak engagement per analyst
      responses:
        '200':
          description: Days-to-peak metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DaysToPeakResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/analytics/team:
    get:
      tags:
        - Analytics
      summary: Get team analytics for all analysts
      responses:
        '200':
          description: Analyst metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsTeamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/images/generate:
    post:
      tags:
        - Images
      summary: Generate an image concept from prompt text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateImageRequest'
      responses:
        '200':
          description: Image concept created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedImageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/images/generate-for-draft:
    post:
      tags:
        - Images
      summary: Generate an image concept for an existing draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateImageForDraftRequest'
      responses:
        '200':
          description: Image concept created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedImageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/images/for-draft/{draftId}:
    get:
      tags:
        - Images
      summary: List generated images for a draft
      parameters:
        - $ref: '#/components/parameters/DraftIdParam'
      responses:
        '200':
          description: Generated images
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedImagesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DraftIdParam:
      name: draftId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadGateway:
      description: Upstream AI/provider request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - error
        - timestamp
      properties:
        ok:
          type: boolean
          enum:
            - false
        error:
          type: string
        details:
          nullable: true
          description: Additional validation or provider-specific detail payload.
        timestamp:
          type: string
          format: date-time
    SuccessEnvelope:
      type: object
      required:
        - ok
        - timestamp
      properties:
        ok:
          type: boolean
          enum:
            - true
        timestamp:
          type: string
          format: date-time
...
    AnalyticsType:
      type: string
      enum:
        - DRAFT_CREATED
        - DRAFT_POSTED
        - FEEDBACK_GIVEN
        - VOICE_REFINEMENT
        - REPORT_INGESTED
        - ENGAGEMENT_RECORDED
        - SESSION_START
        - RESEARCH_CONDUCTED
        - ALERT_GENERATED
        - IMAGE_GENERATED
...
    ImageStyle:
      type: string
      enum:
        - infographic
        - quote_card
        - avatar
        - thumbnail
...
    UserProfile:
      type: object
      required:
        - id
        - handle
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        supabaseId:
          type: string
          nullable: true
        handle:
          type: string
        email:
          type: string
          format: email
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
        onboardingTrack:
          allOf:
            - $ref: '#/components/schemas/OnboardingTrack'
          nullable: true
        telegramChatId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        voiceProfile:
          allOf:
            - $ref: '#/components/schemas/VoiceProfile'
          nullable: true
...
    AnalyticsTeamCount:
      type: object
      required:
        - tweetDrafts
        - analyticsEvents
        - sessions
      properties:
        tweetDrafts:
          type: integer
        analyticsEvents:
          type: integer
        sessions:
          type: integer
...
    AnalyticsTeamUser:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          required:
            - _count
          properties:
            _count:
              $ref: '#/components/schemas/AnalyticsTeamCount'
...
    AnalyticsEvent:
      type: object
      required:
        - id
        - userId
        - type
        - createdAt
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          $ref: '#/components/schemas/AnalyticsType'
        value:
          type: number
          format: float
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
    LearningLogEntry:
      type: object
      required:
        - id
        - userId
        - event
        - impact
        - positive
        - createdAt
      properties:
        id:
          type: string
        userId:
          type: string
        event:
          type: string
        impact:
          type: string
        positive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AnalyticsSummary:
      type: object
      required:
        - draftsCreated
        - draftsPosted
        - feedbackGiven
        - refinements
        - reportsIngested
        - period
      properties:
        draftsCreated:
          type: integer
        draftsPosted:
          type: integer
        feedbackGiven:
          type: integer
        refinements:
          type: integer
        reportsIngested:
          type: integer
        period:
          type: string
    EngagementDailyPoint:
      type: object
      required:
        - date
        - dayLabel
        - predicted
      properties:
        date:
          type: string
          format: date
        dayLabel:
          type: string
        predicted:
          type: integer
        actual:
          type: integer
          nullable: true
    ActivityDailyPoint:
      type: object
      required:
        - date
        - count
      properties:
        date:
          type: string
          format: date
        count:
          type: integer
    TeamEngagementDailyPoint:
      type: object
      required:
        - date
        - dayLabel
        - modelTarget
        - teamActual
      properties:
        date:
          type: string
          format: date
        dayLabel:
          type: string
        modelTarget:
          type: integer
        teamActual:
          type: integer
    PeakRecord:
      type: object
      required:
        - name
        - days
        - hasDrafts
      properties:
        name:
          type: string
        days:
          type: integer
        hasDrafts:
          type: boolean
...
    GeneratedImage:
      type: object
      required:
        - id
        - userId
        - prompt
        - style
        - imageUrl
        - mimeType
      properties:
        id:
          type: string
        userId:
          type: string
        draftId:
          type: string
          nullable: true
        prompt:
          type: string
        style:
          $ref: '#/components/schemas/ImageStyle'
        imageUrl:
          type: string
        mimeType:
          type: string
        createdAt:
          type: string
          format: date-time
    GeneratedImageWithConcept:
      allOf:
        - $ref: '#/components/schemas/GeneratedImage'
        - type: object
          required:
            - concept
          properties:
            concept:
              $ref: '#/components/schemas/VisualConcept'
...
    CreateLearningLogRequest:
      type: object
      required:
        - event
        - impact
      properties:
        event:
          type: string
          minLength: 1
        impact:
          type: string
          minLength: 1
        positive:
          type: boolean
          default: true
...
    GenerateImageRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 5000
        style:
          allOf:
            - $ref: '#/components/schemas/ImageStyle'
          default: quote_card
    GenerateImageForDraftRequest:
      type: object
      required:
        - draftId
      properties:
        draftId:
          type: string
          minLength: 1
        style:
          allOf:
            - $ref: '#/components/schemas/ImageStyle'
          default: quote_card
...
    AnalyticsSummaryResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - summary
              properties:
                summary:
                  $ref: '#/components/schemas/AnalyticsSummary'
    LearningLogEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - entry
              properties:
                entry:
                  $ref: '#/components/schemas/LearningLogEntry'
    LearningLogListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - entries
              properties:
                entries:
                  type: array
                  items:
                    $ref: '#/components/schemas/LearningLogEntry'
    AnalyticsEventsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - events
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/AnalyticsEvent'
    EngagementDailyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - days
              properties:
                days:
                  type: array
                  items:
                    $ref: '#/components/schemas/EngagementDailyPoint'
    ActivityDailyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - days
              properties:
                days:
                  type: array
                  items:
                    $ref: '#/components/schemas/ActivityDailyPoint'
    TeamEngagementDailyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - days
              properties:
                days:
                  type: array
                  items:
                    $ref: '#/components/schemas/TeamEngagementDailyPoint'
    DaysToPeakResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - peaks
              properties:
                peaks:
                  type: array
                  items:
                    $ref: '#/components/schemas/PeakRecord'
    AnalyticsTeamResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - analysts
              properties:
                analysts:
                  type: array
                  items:
                    $ref: '#/components/schemas/AnalyticsTeamUser'
...
    GeneratedImageEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - image
              properties:
                image:
                  $ref: '#/components/schemas/GeneratedImageWithConcept'
    GeneratedImagesResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - images
              properties:
                images:
                  type: array
                  items:
                    $ref: '#/components/schemas/GeneratedImage'
...